Split base test fixture: common utilities base and test case base.
Part of preparation to kickstart writing BDD-style tests
which will re-use the base_test_fixture with common utilities
in tests based on Catch's SCENARIO_METHOD,
while the test_case_fixture will remain solely used by tests
based on Catch's TEST_CASE_METHOD.
